core: Mark classes and functions with _HL_API
Add a new header hl-exports.h, which defines a new macro _HL_API, which can be
used to export symbols where automatic exporting by the compiler/linker is not
supported unless explicitly marked.

Also, for Visual-Studio-style compilers, define _HL_API as appropriate so that
the symbols can be properly exported when building highlight-core as a DLL, and
properly imported when being used to build the highlight-cli program and the QT
GUI program.